+++ src/academy/everyonecodes/java/week2/reflection/exercise1/Product.java
@@ -2,9 +2,9 @@ package academy.everyonecodes.java.week2.reflection.exercise1;
 
 public class Product {
     private String name;
-    private int price;
+    private double price;
 
-    public Product(String name, int price) {
+    public Product(String name, double price) {
         this.name = name;
         this.price = price;
     }
@@ -13,11 +13,11 @@ public class Product {
         return name;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 }
